Add Move Command
Also fix some compiler warnings. 280/284 tests passing, 4 tests fail because an exception is thrown when update_geometry is called on the Move_command result.

CGAL error: precondition violation!
Expression : ( i >= 0 && i < 4 ) && ( j >= 0 && j < 4 ) && ( i != j )
File       : /Users/adam/vcpkg/installed/x64-osx/include/CGAL/Triangulation_utils_3.h
Line       : 91
Explanation: 
Refer to the bug-reporting instructions at https://www.cgal.org/bug_report.html

+++ include/Utilities.hpp
@@ -407,8 +407,7 @@ template <typename FloatingPointType>
               << " timeslices desired.\n";
   }
 
-  auto const simplices_per_timeslice =
-      static_cast<int_fast64_t>(simplices / timeslices);
+  auto const simplices_per_timeslice = simplices / timeslices;
   switch (dimension)
   {
     case 3: {
